Allow to insert a cross-reference via dialog directly to target (#1624)
Next to the list of existing labels, the crossref dialog now also allows
to select a target directly from the TOC, list of figures, etc.

If this element already has a label, we use it. Otherwise, we auto-create
a new one and use that (as already possible via the outliner).