Move fixBiblio to updateBuffer (2/4) : rewrite partly fixBiblio
Now that function takes care of cursor update using the newly
introduced updateAfterInsert. This will be changed later to handle all
views of a given document.

As a consequence, do not return an int anymore.

When two bibinsets are present, only the first one is kept. When an
inset needs to be moved to position 0, always use a copy by simplicity
(the old code was wrong when tracking changes but still deleting
elements because they are of same author).