Integrate theorem definition in layout
Rather than using verbatim \newtheorem definitions in the Preamble
tag, introduce dedicated layout tags from which the definition
will be set up.

This prevents redundant code, but more importantly, it is required
for support of theorems via cleveref and zref-clever. They support
many theorems by default, but not the way we used them (see #13176).

Some support is implemented with this commit, some still needs to be
done.

Layout format change.