Improve input for maxima
Maxima and Maple share many constructs and it is probably for
this reason that sometimes a MapleStream is used instead of a
MaximaStream. However this has some unfortunate consequences
because in some cases the output from maxima would depend on
the particular way an input is built. For example, using maxima,
$\sin(\pi)/1=0$ but $\frac{\sin(\pi)}{1}=\sin\pi$. See also the
example file attached to #13178 for other more complex examples.

By adding a boolean member to MapleStream we can discriminate
when the input is meant for Maxima, instead, and act consequently.