Remove spaces after '%' when input is for maxima
Maxima prefixes with the %-sign many mathematical constants.
LyX can do this only for \pi and not for the imaginary unit 'i'
or the Euler's number 'e' because there is no way to distinguish
such constants from mere variables. In mathed one can prefix
those constants with a %-sign but LyX adds a space after it,
preventing Maxima from recognizing them.