Do not use a version of dispatch() that use a DispatchResult if not using it
There are basically two types of dispatch functions. Those which have
no DispatchResult parameter clean up after themselves and do whatever
updates have been requested.

On the other hand, those with a DispatchResult prameter expect to be
part of a larger dispatch() chain and modify the DR that is passed to
them so that it is taken into account.

tl;dr: never declare a DR variable if you do not intend to do
something with it after calling dispatch.