Remove dubious branch example from UserGuide
The main reason for removing it is that it breaks previews. Since it
is a weird example, it seems easier to remove it than to make it work.

Note that, since change tracking is on, the document still break
previews for now.

(cherry picked from commit a3e7819856fdefb1cbae2a021d61850d48b4d8c6)