Amend 05fa236829
If the document class (not only a module) provides arguments
for lists, we handle them.

Since we also account for special delims in the beamer case,
fall though to that in such cases and limit the special handling
to the case where a list gets arguments only by a module (such
as enumitem).