Fix crashe related to simple search and metrics update
These crashes are a consequence of commit 7f85024f, which improved
performance by not caching position of insets not visible on screen.
This makes a difference, for example, in the case of a tabular inset
which has only its top visible, not its cells.

To this end, introduce the helper methods Inset::hasDim() and
Inset::hasGeometry(), which rely on CoordCache.

The later is used in InsetTabular: the methods hitSelectRow and
hitSelectColumn return false early if no geometry information exists
for the inset.