Rerun latex if necessary when generating previews
Recent versions of dvipng may produce a last page whose content
simply states that "LaTeX was unable to guess the total number
of pages correctly. As there was some unprocessed data that
should have been added to the final page this extra page has
been added to receive it. If you rerun the document (without
altering it) this surplus page will go away, because LaTeX now
knows how many pages to expect for this document."

This confuses the lyxpreview2bitmap.py script, making it unable
to distinguish pages that should follow the dvipng route or the
legacy route. However, the presence of the extra page is reported
in the log file and hence we know we have to rerun latex.

(cherry picked from commit 1888d50eb18277fd93aab2859233887ae5919ebf)