Add new Bullet constructor
The default one cannot be relied on because Bullet::operator= is defined.

Bascally, the only reason why this is needed is the presence of calls
to testInvariant. It may be relevant to evaluate the usefulness of
these assertions.

Avoids the following gcc warning:

../../master/src/BufferParams.cpp:340:21: warning: implicitly-declared ‘lyx::Bullet::Bullet(const lyx::Bullet&)’ is deprecated [-Wdeprecated-copy]
  340 | class BufferParams::Impl
      |                     ^~~~
In file included from ../../master/src/BufferParams.cpp:24:
../../master/src/Bullet.h:53:18: note: because ‘lyx::Bullet’ has user-provided ‘lyx::Bullet& lyx::Bullet::operator=(const lyx::Bullet&)’
   53 |         Bullet & operator=(Bullet const &);
      |                  ^~~~~~~~