lyxbuild: enable compilation with '-pedantic'
'-pedantic' warnings are not too useful, but solving them may make
compilation smoother for compilers that LyX developers are not
testing.

I tested and it works well with both GCC and Clang. That is, it
compiles successfully now. e.g., if 231b45a0 is reverted (which
fixed a -pedantic warning), compilation with lyxbuild fails with
both GCC and Clang.

-Wnested-anon-types is excluded since it is not clear that fixing
those is worth the time.