Use macros that were not meant to be used
The macros UpdateLocker and PathChanger were introduced as a way to
catch incorrect usage of the corresponding classes.

e.g., the intended way to use UpdateLocker is:

  UpdateLocker locker(...);

The incorrect way to use it is:

  UpdateLocker(...);

The macro protects against the incorrect usage.

Thus, it is defined but not intended to be used in our code.
However, this triggered warnings about unused macros. This commit
then uses the macros in a no-op way. This approach is documented
here:

  https://gcc.gnu.org/onlinedocs/gcc/Warning-Options.html

This commit fixes warnings from Clang and GCC. Here are the GCC
warnings:

  src/support/PathChanger.cpp:65: warning: macro "PathChanger" is not used [-Wunused-macros]

  src/mathed/InsetMathMacro.cpp:682: warning: macro "UpdateLocker" is not used [-Wunused-macros]