Allow including the symbols system file from the user one
Until now, in order to make additions to the lib/symbols file
one should have copied it to the user dir and modify the copy.
Of course, when the system file is updated one should transfer
the modifications to the copy. Of course, this is tedious and
error prone. This commit introduces the possibility of including
the system file from the one in the user directory by using the
directive "include_system_file", which is equivalent to inserting
the system symbols file at that point. The, one can make additions
or modifications as desired.