Correctly parse labels enclosed in braces in mathed
If a label in math is enclosed in braces, the math parser
is not able to correctly assign the label to the GUI inset.
Hence, references to these labels are marked as broken in
the GUI, even if the document could still compile correctly.
Howerver, in these cases the label is parsed as a regular math
construct, so that, for example, "a_b" becomes a script inset
and then the argument of the label is exported as "a_{b}" and
references to "a_b" cause compilation errors.
After this commit the argument of the label is stored verbatim
and can thus be recovered and correctly shown in the GUI.
Fixes #13195.