tex2lyx: more clever roman font package detection
If two roman font package are loaded in a row, this does not
necessarily mean the latter completely supersedes the former.

E.g.,

\usepackage{mathpazo}
\usepackage{charter}

still uses the pazo fonts for math. If we simply overwrite
the roman font, this gets lost. So we need to keep the preamble
code in this case.