Fix preview drawing after mouse selection is finished
In order to fix #8951, commit 22030781 was backported from gadmm's
lyx-unstable branch. What this does is record when one is selecting,
and block the the switch to preview when the cursor tip leaves the
previewable inset.

The problem is that there is no notification given when the selection
action is finished (mouse release) and therefore metrics are not
recomputed. So the equation is correctly shown as a preview but relies
on the old no-preview metrics.

This commit implements such a notification:

- Add virtual method notifyMouseSelectionDone() to Inset, InsetMathHull,
  InsetPreview and InsetIPA.

- Add function notifyMouseSelectionDone() to Cursor.cpp.

- In BufferView::mouseEventDispatch, call the function above when
  needed. Note that the code has been changed to preserve the cursor's
  anchor. There might be code that does not expect that.

Fixes bug #13222.